+++ src/system.h
@@ -3,4 +3,4 @@
 **    system.h                        NQ                       Werner Nickel
 */
 
-extern long RunTime (); 
+extern int RunTime ();
